ref(crons): Remove issue_platform module
This has been (mostly) been replaced with detectors
